add support for evaluating expressions
This change allows immediate values to be calculated from an expression,
such as 1+1, even including symbols such as "100 << const" (where const
was defined with the .set directive). Expressions are also supported in
the .set directives.

Expressions are evaluated using the built-in eval(). To prevent misuse or
malicious code execution, expressions are validated.

At the point when eval is called, all symbols should have already been resolved
to their values. That means we only need to allow for numeric characters along
with arithmetic and bitwise operators, round brackets and whitespace. The
character 'x' and the characters 'abcdef' are also accepted to allow for hex
numbers such as 0x123abc. These are only allowed however in sequences starting
with 0x. If any other character is encountered the expression is deemed invalid
and an exception is raised.