add a compatibility test for evaluating expressions
This is a common example, where the ADC is read multiple times (oversampled)
and the oversampling factor is calculated by a shift-left, which makes
calculating the average of samples as easy as shifting their sum right by
the same amount of bits. Using .set directives with expressions makes
the oversampling factor easily configurable.