add preprocessor that can replace simple #define values in code
The preprocessor strips all comments and lines containing a define statement
but keeps the empty lines in the output (to preserve line numbering). The
output is then passed directly into the assembler.

The preprocessor does not support "function style" #define macros (i.e.
ADD(a,b) a+b) but this is not needed for expanding the constants used by
WRITE_RTC_REG(), et al.