add support for macros such as WRITE_RTC_REG
This is a simplified implementation, rather than adding "proper" support for macros
because we don't need more. The macros WRITE_RTC_REG, READ_RTC_REG, WRITE_RTC_FIELD
and READ_RTC_FIELD are simply expanded in a predefined way. If they are also defined
as macros in the source code, those macros in the source will be ignored.