allow specifying the address for reg_rd and reg_wr in 32-bit words
This change allows specifying the address in 32-bit words (i.e. the
address as seen from the ULP), in addition to the existing mode of
specifying a register's full address on the DPORT bus.

If an address is between 0 and DR_REG_MAX_DIRECT (0x3ff), treat it
as a word offset (ULP address), otherwise treat it as a full address
on the DPORT bus as before.