allow assembler to skip comment removal to avoid removing comments twice
Since the preprocessor was introduced, which already removes comments, the
assembler does not need to remove comments anymore in the usual case. The
assembler still retains the ability to remove comments (enabled by default)
in case it is used without the preprocessor. The `remove_comments` argument
to the `assemble()` method can be used to control whether comments will be
removed during assembly or not.