add compat tests for relative jumps with immediate offset from absolute symbol
Values assigned to symbols using .set should be specified in words (not
bytes). Thus, this test ensures that the value of the const symbol is not
converted to words (divided by 4).

Since py-esp32-ulp internally tracks symbol values as words anyway, this
behaviour has been correct all along. The test aims to show that recent
changes related to immediate value handling did not break this (and to
guard against future changes that could break this).