add blink example
This example makes the ULP turn on and off GPIO2 at regular intervals.
If one attaches an LED to GPIO2 and GND, it would blink with a period
of approximately 500ms for on and off each.

The example defines all needed RTC constants, so it is usable even
without a populated Defines DB.

The example ends with a loop showing the contents of two memory fields,
which the ULP code updates. This allows seeing that the ULP is working.
(Can be useful for running the example, even when there is no LED
connected to GPIO2).

Fixes #12.