fix unit tests for running on the ESP32 device
When running unit tests on the ESP32, the current directory is root, while
on the PC (or in the GitHub Actions), the tests/ directory is the current
directory. Unit tests that rely on fixture files fails because of this.
Now test scripts can look for fixture files relative to their own path to
make them independent of the current directory.