output SHA1 for files built during compat tests
This allows for comparing binaries generated on the ESP32 with those generated
on the PC with binutils-esp32ulp.

During testing on the ESP32, given that there is no binutils-esp32ulp, only the
py-esp32-ulp part of the compat tests can run. Calculating the SHA1 for each
resulting file from the compat tests, then allows comparing those hashes with
hashes calculated on the PC.