add setup.py to build upip installable distribution package for pypi
setup.py (setuptools) uses the sdist_upip module (included in this
commit) to generate packages in the correct format for use with upip.

To generate a distribution package, make sure the VERSION is set as
desired in setup.py and then run `python3 setup.py sdist`. The result
will be a tar.gz file in the dist/ subdirectory.