add additional supported platforms
It appears the "platforms" fields it not actually used by any tools but
can be used to communicate to users what platforms the package is meant
to work on. (https://stackoverflow.com/a/34994245)

Including linux and darwin here, because the code runs well on the unix
port of MicroPython, which itself runs well on Linux and macOS (darwin).
In fact most of the testing had been done using the unix port.