Make upip installable (#56)
add setup.py to build upip installable distribution package for pypi

setup.py (setuptools) uses the sdist_upip module (included in this
commit) to generate packages in the correct format for use with upip.

To generate a distribution package, make sure the VERSION is set as
desired in setup.py and then run `python3 setup.py sdist`. The result
will be a tar.gz file in the dist/ subdirectory.

also: add additional supported platforms

It appears the "platforms" fields it not actually used by any tools but
can be used to communicate to users what platforms the package is meant
to work on. (https://stackoverflow.com/a/34994245)

Including linux and darwin here, because the code runs well on the unix
port of MicroPython, which itself runs well on Linux and macOS (darwin).
In fact most of the testing had been done using the unix port.