add reference to battery-powered applications to advanced usage in README
While py-esp32-ulp could be used on a PC to generate the ULP binary,
which in turn could be uploaded to an ESP32, this workflow can also
be realised using binutils-esp32ulp on the PC.

py-esp32-ulp makes it easy to work directly on the ESP32 to assemble
code for the ULP. However, the ESP32 is rather slow at doing this,
which becomes relevant in battery-powered applications, where every
second of sleep time is valuable.

The documentation therefore no longer refers to the use-case of
assembling ULP code on a PC, and adds a short reference to battery-
powered applications as one reason, why one might want to split the
assembly and load phase and store the ULP binary into a file.