add workflow to publish to PyPI
This workflow is triggered when publishing a new release
using the GitHub Releases functionality. It can also be
triggered manually for testing.

The workflow will always publish to Test PyPI, but will
only publish to the real PyPI, when releasing a new
version using the GitHub Releases functionality.

The version for the package is generated by setuptools_scm,
based on the git tags/commits. For all non-tagged commits,
a "dev version" is generated with the format "0.1.2-dev3",
where 0.1.2 is the latest tag plus one more in the last part
of the version number (i.e. given the last tag was 0.1.1,
0.1.2 will be used for the dev version) and the 3 in "-dev3"
comes from the number of commits since the last tag.
(This all is standard setuptools_scm behaviour.)