update MANIFEST.in to package only the main source
The MANIFEST.in will exclude documentation, examples and tests
and all files in the repo root. This way only the main source
in esp32_ulp/ will be packaged by sdist.

Since we're packaging to upip format, which can only be installed
by MicroPython/upip, we leave out all files that are not strictly
needed to run on the ESP32, so we save space on the device.

If anyone needs the other files on a more capable device (aka PC),
simply do a git clone.