Use micropython from new build location during build
The MicroPython project changed their CI scripts slightly
which resulted in the unix port binary being output to a
subdirectory (build-standard) instead of the current
directory as before. This broke our build process.

This commit updates our build script to search for
the micropython binary in the correct subdirectory.

Fixes #81