Update RTC compat tests to use new binutils-gdb (esp32ulp)
The examples from the testsuite are the same as in the previous
binutils-esp32ulp repo from Espressif. But since the old repo
will no longer be maintained, we switch to the new one.

Currently all development happens on the esp32ulp-elf-2.35 branch
and master does not (yet?) contain any files related to the
Espressif-adapted assembler/linker. So we use this branch for now.

Given that the branch name includes what appears to be a version
number, it could mean Espressif might switch to a different branch
down the line. We should pay attention to this and can then choose
to switch to that branch too.