Explicitly target the ESP32 during binary comparison
When using binutils-gdb (esp32ulp) to generate the "reference binary",
which we compare against, pass --mcpu=esp32 to select the original
ESP32 variant. While this is the default target of esp32-ulp-as,
adding this argument makes the choice explicit and shows how to select
a variant, for when we add support the ESP32-S2 or ESP32-S3 (both
use --mcpu=esp32s2 because their ULPs have the same binary format).