Add hex representation for any field value > 9
Some values are easier to read as hex values than as decimal.
For example peripheral register addresses like 0x123 where the
first digit (1) indicates which peripheral register to address,
while the remaining 2 digits (0x23) are the offset within that
register in number of 32-bit words.

Also absolute JUMP addresses are easier to find via the hex value
given that the disassembler includes the byte offset of each
instruction in hex format.