Allow selecting cpu to disassemble for
Currently only the esp32 is implemented (esp32s2 soon to follow).

This commit adds the -c command line option to the disassembler, as
well as updates the integration tests to supply the cpu parameter,
and test fixtures are renamed to include the cpu type in their name,
so that we can have separate fixture files for other cpus.