Add support for new ST and LD instructions to the disassembler
The following new instructions are now supported:
* LDL, LDH
* STL, STH, ST32, STI32, STI, STO

Note: The disassembler will return LD instead of LDL and
ST instead of STL, because they are each synonyms of the
other. We can only pick either and so we picked the keyword
that exists across both the ESP32 and the ESP32-S2/S3.