Support peripheral register addresses of ESP32-S2/S3
Peripheral registers of the ESP32-S2 and S3 are at a different
location than those of the original ESP32. The location within
the address space is mostly the same, however we need to use
the correct base address when calculating periph_sel (type of
register) used in REG_RD and REG_WR instructions.

Note 1: To avoid creating an entirely new CPU (esp32s3) just for
handling the different peripheral register addresses of the S3,
while their binary format (instruction set) is identical, our
esp32s2 CPU support will now accept both ESP32-S2 and ESP32-S3
addresses. This should make a binary for the one seamlessly
work on the other (without reassembly), given that the offsets
of different peripheral registers between the S2 and S3 are
mostly (but not entirely) identical.

Note 2: Our esp32s2 cpu support will also accept peripheral
register addresses of the original ESP32. This was originally
done because Espressif's binutils-gdb/esp32-ulp-as incorrectly
validates addresses for the esp32s2 cpu, and to make our compat
tests pass, this was needed. However this also has a nice side-
effect of allowing some assembly written for the original ESP32
to work unmodified when assembled for an S2/S3, because some of
the peripheral registers live at the same offset from the base
for all three variants.