Add examples for ESP32-S2 and ESP32-S3
There are now example files for the S2 and S3 (with ``_s2`` or ``_s3``
appended to their filenames).

Note: The s2 examples also work unmodified on the ESP32-S3, except the
readgpio example which needs different peripheral register addresses
on the S3.

The ``counter_s2.py`` example is unmodified compared to the original
example, except that the assembler is told to generate esp32s2 output.

The ``blink_s2.py``, ``readgpio_s2.py`` and ``readgpio_s3.py`` examples
have their rtc_io base address updated, as well as the constants
referring to the GPIO pins and channels and the peripheral register bits
used to read/write the GPIO inputs/outputs. These addresses/bits have
changed from the original ESP32. Otherwise the examples are identical to
the examples for the original ESP32.