Add licence headers to all source files
The licence header explains that the given file is part of this
project, and expresses the copyright information and the licence
the file is licensed under using SPDX headers to keep the header
short.

While we have a repo-global LICENSE file, having licence text
directly in a source file makes it very clear for anyone who
copies the file what the licence/copyright is.